Incorporated all review comments
